Column tests refactoring first attempt
